fix circular dependency warnings
